Not silently swallowing error if server side render/mount fails
If the app crashes on a runtime server render, specifically at the
ReactDOM.renderToString call, the .then() will silently swallow the
error and the request will be left spinning until timeout. The error
will not be logged/shown *anywhere* so debugging is impossible. Adding a
catch() with logging exposes the error to the server logs, and 500s.

Maybe worth adding https://github.com/xjamundx/eslint-plugin-promise ?